requisito => 04- concluido com sucesso

+++ src/models/MongoModel.ts
@@ -1,4 +1,4 @@
-import { Model } from 'mongoose';
+import { isValidObjectId, Model, UpdateQuery } from 'mongoose';
 import { IModel } from '../interfaces/IModel';
 
 abstract class MongoModel <T> implements IModel<T> {
@@ -13,14 +13,22 @@ abstract class MongoModel <T> implements IModel<T> {
   }
 
   public async readOne(_id: string): Promise<T | null> {
+    if (!isValidObjectId(_id)) throw new Error('Erro Read one model');
     return this.model.findOne({ _id });
   }
 
   public async update(_id: string, obj: T): Promise<T | null> {
-    return this.model.findOneAndUpdate({ _id }, { obj });
+    if (!isValidObjectId(_id)) throw Error('Erro update model');
+
+    return this.model.findOneAndUpdate(
+      { _id },
+      { ...obj } as UpdateQuery <T>,
+      { new: true },
+    );
   }
 
   public async delete(_id: string): Promise<T | null> {
+    if (!isValidObjectId(_id)) throw Error('Erro delete model');
     return this.model.findOneAndDelete({ _id });
   }
 }
